BookService and book search
BookService: service injected into components to load the book data. Currently the books are hardcoded to the service, but in future commits it will fetch the data from the Library API
MainList: Component with a BookList and a text input to filter books by title. Receives book data from BookService and passes it to BookList